// src/features/dashboard/types/filterInputType.ts
export type FilterInput = {
  adjustment?: number;
  bedLength?: string;
  bodyStyle?: string;
  doors?: string;
  drivetrain?: string;
  engine?: string;
  engineCylinders?: string;
  engineFuelType?: string;
  engineSizeLitters?: number;
  make?: string;
  model?: string;
  transmission?: string;
  trim?: string;
  year?: number;
};